add start of field processing in discover/aws
Starting slow here. We add the initial processing of fields for a
resource in the `pkg/discover/aws` package. For now, not handling field
definitions (typing), only handling field names and paths along with a
facility to rename a field and search for a FieldConfig by field path,
field name or renamed component.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>